backend/requirements.txt
fastapi
uvicorn==0.30.0
bleach==6.2.0
certifi==2024.12.14
charset-normalizer==3.4.0
filelock==3.16.1
fsspec==2024.10.0
idna==3.10
Jinja2==3.1.4
kaggle==1.6.17
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.0
pillow==11.0.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
requests==2.32.3
setuptools==75.6.0
six==1.17.0
sympy==1.13.1
text-unidecode==1.3
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
webencodings==0.5.1
Pillow==11.0.0
scikit-learn
matplotlib
requests
python-dotenv
pandas
flask
flask-cors